Avoid split runtime packages on MongoDB with Panache
(cherry picked from commit 11899b588cab2dee0f59ccfbf2d2ebab4abf133e)
