+++ extensions/panache/mongodb-panache/runtime/src/main/java/io/quarkus/mongodb/panache/runtime/JavaMongoOperations.java
@@ -10,6 +10,8 @@ import com.mongodb.client.MongoCollection;
 
 import io.quarkus.mongodb.panache.PanacheQuery;
 import io.quarkus.mongodb.panache.PanacheUpdate;
+import io.quarkus.mongodb.panache.common.runtime.MongoOperations;
+import io.quarkus.mongodb.panache.common.runtime.PanacheUpdateImpl;
 
 public class JavaMongoOperations extends MongoOperations<PanacheQuery<?>, PanacheUpdate> {
     /**
